move: `@/infrastructure/repositories` -> `@/infrastructure/strapi` ファイル名も変更
